refactor: share file metadata query
